- Se organiza el proceso para evitar recibir medios de pago menores o iguales a 0
- Se organiza proceso para generar codigos de barra
- Se organiza problema al validar el stock
- Inicia la implementacion de cartera
- Solucionado busqueda por color y talla
- Se soluciona error al validar inventarios
- Se organiza la consulta de usuarios

+++ src/inventories/services/stock-adjustment.service.ts
@@ -263,16 +263,21 @@ export class StockAdjustmentService {
 				documentId: response._id.toString(),
 				documentType: DocumentTypeStockHistory.ADJUSTMENT,
 			};
-			await this.stockHistoryService.deleteStock(
-				deleteStockHistoryInput,
-				user,
-				companyId,
-			);
-			await this.stockHistoryService.addStock(
-				addStockHistoryInput,
-				user,
-				companyId,
-			);
+
+			if (deleteStockHistoryInput?.details?.length > 0) {
+				await this.stockHistoryService.deleteStock(
+					deleteStockHistoryInput,
+					user,
+					companyId,
+				);
+			}
+			if (addStockHistoryInput?.details?.length > 0) {
+				await this.stockHistoryService.addStock(
+					addStockHistoryInput,
+					user,
+					companyId,
+				);
+			}
 		}
 		return response;
 	}
@@ -430,16 +435,22 @@ export class StockAdjustmentService {
 					documentId: response._id.toString(),
 					documentType: DocumentTypeStockHistory.ADJUSTMENT,
 				};
-				await this.stockHistoryService.deleteStock(
-					deleteStockHistoryInput,
-					user,
-					companyId,
-				);
-				await this.stockHistoryService.addStock(
-					addStockHistoryInput,
-					user,
-					companyId,
-				);
+
+				if (deleteStockHistoryInput?.details?.length > 0) {
+					await this.stockHistoryService.deleteStock(
+						deleteStockHistoryInput,
+						user,
+						companyId,
+					);
+				}
+
+				if (addStockHistoryInput?.details?.length > 0) {
+					await this.stockHistoryService.addStock(
+						addStockHistoryInput,
+						user,
+						companyId,
+					);
+				}
 			}
 
 			return response;
@@ -508,16 +519,21 @@ export class StockAdjustmentService {
 					documentType: DocumentTypeStockHistory.ADJUSTMENT,
 				};
 
-				await this.stockHistoryService.deleteStock(
-					deleteStockHistoryInput,
-					user,
-					companyId,
-				);
-				await this.stockHistoryService.addStock(
-					addStockHistoryInput,
-					user,
-					companyId,
-				);
+				if (deleteStockHistoryInput?.details?.length > 0) {
+					await this.stockHistoryService.deleteStock(
+						deleteStockHistoryInput,
+						user,
+						companyId,
+					);
+				}
+
+				if (addStockHistoryInput?.details?.length > 0) {
+					await this.stockHistoryService.addStock(
+						addStockHistoryInput,
+						user,
+						companyId,
+					);
+				}
 			}
 
 			return response;
